Readding from scratch instead reverting the revert.
It kept eating the files I added, so I recreated
the changelist from scratch using the raw patch 
downloaded from Rietveld.

TBR=mpcomplete

Original description:

We changed the page action manifest from
 'icon' -> 'default_icon'
and now allow extensions to programmatically
specifying an icon. However, page actions that
hard-coded an icon using the 'icon' tag in the
manifest (don't show an icon programmatically)
now lose their icons. This causes a crash when we
try to display the icon in the LocationBarView.

We now check when the extension tries to enable
the page action whether there are any icons to
display. If not, we don't proceed and log an
error to the console.

BUG=25562
TEST=Covered by browser test.

Review URL: http://codereview.chromium.org/334005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@29875 0039d316-1c4b-4281-b951-d872f2087c98