gtk: Show page action popup bubbles.
Focus is a bit weird within these; I suspect that it's the same
with the browser action bubbles.  It looks like the same issue that
the bookmark bubble had: the bubble grabs the pointer; when a
widget within it (e.g. combobox) gets clicked it steals the grab;
when it ungrabs it the grab is released instead of being
transferred back to the bubble.  Unsure of what the correct fix
is for this.

BUG=28061
TEST=installed two separate extensions with page action popups and clicked on their icons

Review URL: http://codereview.chromium.org/412001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@32547 0039d316-1c4b-4281-b951-d872f2087c98