Revert 36589 - Still failing the reliability tests.
Another try at landing this.
It caused several tests to fails previously.
There was a crasher in some Windows code in the CLD that has been removed by the port of the CLD code to Linux and Mac.
That should hopefully make everything work now.

Enabling language detection on page load.
A memory error has been fixed in the CLD library in the meantime.
This should hopefully fixes the crashers in the reliability tests.
Note that this version is actually simpler than the original review since the detection is now performed in the renderer. (So the CLD code runs sandboxed.)

Original review:
http://codereview.chromium.org/492024/show

BUG=30662
TEST=Run the unittests.
TBR=brettw

Review URL: http://codereview.chromium.org/552049

TBR=jcampan@chromium.org
Review URL: http://codereview.chromium.org/554015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@36603 0039d316-1c4b-4281-b951-d872f2087c98