Test cookie prompting.
Adds automation plumbing to support changing the content settings.

This change also fixes WaitForNavigation to watch for the LOAD_STOP event
instead of the NAV_ENTRY_COMMITTED event.  This better matches the intent (I
think) of the automation interface, and it also matches the similar function
defined in ui_test_utils (used by browser tests).  This also better matches the
NavigationNotificationObserver, which seems desirable given the implementation
of AutomationProvider::WaitForNavigation.

It turns out that only one UI test was using WaitForNavigation, and I confirmed
that it still functions properly.

This change also modifies the NavigationNotificationObserver to allow it to
watch for the current page load to complete.  Normally, it waits for the first
LOAD_START or NAV_ENTRY_COMMITTED event to switch into the mode where it
watches for LOAD_STOP.  However, that makes it not so useful for
WaitForNavigation, which is intended to be called after triggering the start of
a new navigation.

This change also makes it so that the cookie prompt supports the automation
message box 'OK' and 'Cancel' buttons, so that we can remotely press those
buttons.

Leveraging all of these changes enables some basic cookie prompting tests to be
written.

I had to modify browser_frame_win.cc to ensure that BrowserList::SetLastActive
is called even while running under remote desktop.

R=phajdan.jr@chromium.org
BUG=37182
TEST=cookie_modal_dialog_uitest.cc

Review URL: http://codereview.chromium.org/661102

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@40575 0039d316-1c4b-4281-b951-d872f2087c98