Fix a regression introduced in r40465.
In DoSSLConnectComplete, we still need to set using_spdy_
if we get a certificate error.

In HandleCertificateError, we need to ignore the error
after adding the certificate to the allowed_bad_certs list.

R=mbelshe
BUG=37367
TEST=Connect to a SPDY server with a bad certificate. Clicking
"Proceed anyway" in the SSL blocking page should not result in
a download.
Review URL: http://codereview.chromium.org/668111

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@40687 0039d316-1c4b-4281-b951-d872f2087c98