linux: make seccomp sandbox also use OOP font loading
Within the SUID sandbox, we IPC up to the browser process when
we need to open font files.  We should do the same thing within
the seccomp sandbox.  (Note that in test_shell and in SELinux,
which uses finer grained file access permissions, we don't
need to do such a thing.)

BUG=26522

Review URL: http://codereview.chromium.org/1541002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@42986 0039d316-1c4b-4281-b951-d872f2087c98