[Mac] Move star button into page-actions area of omnibox.
Removes the star button from the toolbar entirely.  Adds a
LocationBarImageView subclass for the star icon and wires it to the
RHS of the field.  Adjust the bookmark bubble to move the arrow to the
RHS and position appropriately.

BookmarkBubble.xib: outlet to view so that controller can change the arrow from left to right.

Toolbar.xib: Move reload icon to where star was, remove star icon, adjust spacing.  All spacing was adjusted to specific positions in the relevant inspector, not by mouse drag, so hopefully there's nothing awry there.

BUG=37865
TEST=No star icon on toolbar.
TEST=Optional home button adjustments should work right.
TEST=Star action in omnibox when showing an URL.
TEST=Star action can be clicked to bookmark current page.
TEST=Bookmark bubble arrow points at star like before.
TEST=Command-d brings up bookmark bubble.
TEST=Star action changes from blank to yellow depending on state.
TEST=Star action tooltip changes depending on state.

Review URL: http://codereview.chromium.org/1540009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@43376 0039d316-1c4b-4281-b951-d872f2087c98